Extract tool observation handling
